refactored to put PostService in Services package
